change made

+++ src/events/events.controller.ts
@@ -13,6 +13,11 @@ import {
     NotFoundException,
     Query,
     UsePipes,
+    UseGuards,
+    ForbiddenException,
+    SerializeOptions,
+    UseInterceptors,
+    ClassSerializerInterceptor,
 } from '@nestjs/common';
 import { CreateEventDto } from './input/create-event.dto';
 import { UpdateEventDto } from './input/update-event.dto';
@@ -22,8 +27,12 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Attendee } from './attendee.entity';
 import { EventsService } from './events.service';
 import { ListEvents } from './input/list.events';
+import { CurrentUser } from 'src/auth/current-user.decorator';
+import { User } from 'src/auth/user.entity';
+import { AuthGuardJwt } from 'src/auth/auth-guard.jwt';
 
 @Controller('/events')
+@SerializeOptions({ strategy: 'excludeAll' })
 export class EventsController {
     private readonly logger = new Logger(EventsController.name);
     constructor(
@@ -36,6 +45,7 @@ export class EventsController {
 
     @Get()
     @UsePipes(new ValidationPipe({ transform: true }))
+    @UseInterceptors(ClassSerializerInterceptor)
     async findAll(@Query() filter: ListEvents) {
         this.logger.debug(filter.page);
         this.logger.log('Hit the findAll route');
@@ -53,40 +63,41 @@ export class EventsController {
     }
 
     // @Param() gives the whole params object and @Param('id') gives the id property of the params object
-    @Get('/practice')
-    async practice() {
-        return await this.repository.find({
-            where: [
-                {
-                    id: MoreThan(3),
-                    when: MoreThan(new Date('2020-05-01')),
-                    description: Like('%meet%'),
-                },
-            ],
-            take: 2,
-            order: {
-                id: 'DESC',
-            },
-        });
-    }
+    // @Get('/practice')
+    // async practice() {
+    //     return await this.repository.find({
+    //         where: [
+    //             {
+    //                 id: MoreThan(3),
+    //                 when: MoreThan(new Date('2020-05-01')),
+    //                 description: Like('%meet%'),
+    //             },
+    //         ],
+    //         take: 2,
+    //         order: {
+    //             id: 'DESC',
+    //         },
+    //     });
+    // }
 
-    @Get('/practice2')
-    async practice2() {
-        // return await this.repository.findOne({
-        //     where: { id: 2 },
-        //     loadEagerRelations: false,
-        // });
-        const event = await this.repository.findOne({
-            where: { id: 1 },
-        });
-        const attendee = new Attendee();
-        attendee.name = 'Jerry';
-        attendee.event = event;
-        await this.attendeeRepository.save(attendee);
-        return attendee;
-    }
+    // @Get('/practice2')
+    // async practice2() {
+    //     // return await this.repository.findOne({
+    //     //     where: { id: 2 },
+    //     //     loadEagerRelations: false,
+    //     // });
+    //     const event = await this.repository.findOne({
+    //         where: { id: 1 },
+    //     });
+    //     const attendee = new Attendee();
+    //     attendee.name = 'Jerry';
+    //     attendee.event = event;
+    //     await this.attendeeRepository.save(attendee);
+    //     return attendee;
+    // }
 
     @Get(':id')
+    @UseInterceptors(ClassSerializerInterceptor)
     async findOne(@Param('id', ParseIntPipe) id: number) {
         // console.log(typeof id);
         // wrap this in try catch block
@@ -105,35 +116,47 @@ export class EventsController {
     }
 
     @Post()
-    async create(
-        @Body(new ValidationPipe({ groups: ['create'] })) input: CreateEventDto,
-    ) {
-        return await this.repository.save({
-            ...input,
-            when: new Date(input.when),
-        });
+    @UseGuards(AuthGuardJwt)
+    @UseInterceptors(ClassSerializerInterceptor)
+    async create(@Body() input: CreateEventDto, @CurrentUser() user: User) {
+        return await this.eventsService.createEvent(input, user);
     }
 
     @Patch(':id')
+    @UseGuards(AuthGuardJwt)
+    @UseInterceptors(ClassSerializerInterceptor)
     async update(
         @Param('id') id,
         @Body(new ValidationPipe({ groups: ['create'] })) input: UpdateEventDto,
+        @CurrentUser() user: User,
     ) {
         const event = await this.repository.findOneBy({ id });
-        return await this.repository.save({
-            ...event,
-            ...input,
-            when: input.when ? new Date(input.when) : event.when,
-        });
+        if (!event) throw new NotFoundException();
+
+        if (event.organizerId !== user.id) {
+            throw new ForbiddenException(
+                null,
+                `You are not authorized to change this event`,
+            );
+        }
+        return await this.eventsService.updateEvent(event, input);
     }
 
     @Delete(':id')
+    @UseGuards(AuthGuardJwt)
     @HttpCode(204)
-    async remove(@Param('id') id) {
-        const result = await this.eventsService.deleteEvent(id);
-        if (result?.affected !== 1) {
-            throw new NotFoundException();
+    async remove(@Param('id') id, @CurrentUser() user: User) {
+        const event = await this.eventsService.getEvent(id);
+        if (!event) throw new NotFoundException();
+
+        if (event.organizerId !== user.id) {
+            throw new ForbiddenException(
+                null,
+                `You are not authorized to remove this event`,
+            );
         }
+
+        await this.eventsService.deleteEvent(id);
     }
 }
